UPDATE CameraKeyBoard y main. Probar camara y sus teclas

+++ APIS_2025/Camera.h
@@ -26,6 +26,6 @@ public:
 	void computeProjectionMatrix();
 	void computeViewMatrix();
 
-	virtual void step(float timestep) = 0;
+	virtual void step(double timestep) = 0;
 };
